Add support for multiple destinations with the same type
For having dual stack configuration, there is a need to specify same
protocol twice, once with IPv4 address and once with IPv6, but currently
doing this results in concat::fragment name conflict.

With this commit, sslh::destination will use title instead of type
variable to make concat::fragment resource title.

This should not affect current users as well, as it's internal
implementation.